refactor print_model_params
